requirements.txt
streamlit
openai>=1.0.0
python-dotenv
sentence-transformers
PyMuPDF
beautifulsoup4
pinecone-client>=3.0.0,<4.0.0
